server/requirements.txt
uvicorn
fastapi
json-numpy
torch
transformers
python-dotenv
draccus
Pillow